chore: Divide up code into modules
Main module was getting way too big.
